Accomplished partial abstraction of the Agent classes themselves.  Nearly all aspects of the simulation handle IAgent objects or List<IAgent> instead of "knowing" about something as specific as AgentSingleIntegerVariant objects.  The places which need to know this cast the agent object to its concrete class right at the point of usage.  This still isn't ideal, because even that cast is a substitute for a generic interface for agent variation in cases where the variant isn't just a single token, but we'll get there.
AbstractAgent is the abstract base class that should be implemented by all future agent classes.  In addition to implementing IAgent, it provides a default implementation for getAgentID() and will do the same for all methods which are common to all agents, so that writing a new agent only has to focus on what is specific to *that* agent class -- e.g. structure of the traits, spatial location, etc.